+++ appcenter/android/src/main/java/com/getcapacitor/plugin/appcenter/AppCenterBase.java
@@ -2,7 +2,6 @@ package com.getcapacitor.plugin.appcenter;
 
 import com.getcapacitor.JSObject;
 import com.microsoft.appcenter.AppCenter;
-import com.microsoft.appcenter.CustomProperties;
 import java.util.Date;
 import java.util.Iterator;
 import java.util.UUID;
@@ -47,37 +46,40 @@ public class AppCenterBase {
         return AppCenter.isEnabled().get();
     }
 
-    public void setCustomProperties(JSObject properties) {
-        CustomProperties customProps = new CustomProperties();
-        Iterator<String> keysIter = properties.keys();
-
-        while (keysIter.hasNext()) {
-            String key = keysIter.next();
-            // JSONObject valueObject = (JSONObject) properties.getJSObject(key);
-            JSObject valueObject = properties.getJSObject(key);
-
-            if (valueObject != null) {
-                String type = valueObject.getString("type");
-                switch (type) {
-                    case "clear":
-                        customProps.clear(key);
-                        break;
-                    case "string":
-                        customProps.set(key, valueObject.getString("value"));
-                        break;
-                    case "number":
-                        customProps.set(key, Double.parseDouble(valueObject.getString("value")));
-                        break;
-                    case "boolean":
-                        customProps.set(key, valueObject.getBool("value"));
-                        break;
-                    case "date-time":
-                        customProps.set(key, new Date((long) Double.parseDouble(valueObject.getString("value"))));
-                        break;
-                }
-            }
-        }
-
-        AppCenter.setCustomProperties(customProps);
+    public void setCountryCode(String countryCode) {
+        AppCenter.setCountryCode(countryCode);
     }
+    // public void setCustomProperties(JSObject properties) {
+    //     CustomProperties customProps = new CustomProperties();
+    //     Iterator<String> keysIter = properties.keys();
+
+    //     while (keysIter.hasNext()) {
+    //         String key = keysIter.next();
+    //         // JSONObject valueObject = (JSONObject) properties.getJSObject(key);
+    //         JSObject valueObject = properties.getJSObject(key);
+
+    //         if (valueObject != null) {
+    //             String type = valueObject.getString("type");
+    //             switch (type) {
+    //                 case "clear":
+    //                     customProps.clear(key);
+    //                     break;
+    //                 case "string":
+    //                     customProps.set(key, valueObject.getString("value"));
+    //                     break;
+    //                 case "number":
+    //                     customProps.set(key, Double.parseDouble(valueObject.getString("value")));
+    //                     break;
+    //                 case "boolean":
+    //                     customProps.set(key, valueObject.getBool("value"));
+    //                     break;
+    //                 case "date-time":
+    //                     customProps.set(key, new Date((long) Double.parseDouble(valueObject.getString("value"))));
+    //                     break;
+    //             }
+    //         }
+    //     }
+
+    //     AppCenter.setCustomProperties(customProps);
+    // }
 }